+++ emails/cleaners.py
@@ -5,7 +5,7 @@ from __future__ import annotations
 from django.contrib.auth.models import User
 from django.db.models import Q
 
-from privaterelay.cleaners import (
+from privaterelay.data_issue_task import (
     CleanerTask,
     DataBisectSpec,
     DataItem,
